added marks function
